fix(typecheck): cast int→uint32 for syscall.Credential.Uid
CI's stricter typecheck (golangci-lint v2.4.0 v8) caught what the
local lint passed through: runAppWithSecrets used uid (int) directly
in syscall.Credential{Uid: uid}, but the struct field is uint32.
runApp (legacy) already casts; mirror that here.

+++ guest/init/main_linux.go
@@ -109,7 +109,7 @@ func runAppWithSecrets(m api.AppManifest, secrets map[string]string) error {
 	cmd.Stdout, cmd.Stderr = os.Stdout, os.Stderr
 	if uid := lookupUID(m.EffectiveUser()); uid > 0 {
 		cmd.SysProcAttr = &syscall.SysProcAttr{
-			Credential: &syscall.Credential{Uid: uid, Gid: uid},
+			Credential: &syscall.Credential{Uid: uint32(uid), Gid: uint32(uid)},
 		}
 	}
 	if err := cmd.Run(); err != nil {
